# api/app/src/swagger/swagger_spec.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1",
        "title": "MethodFitness"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/purchase/purchase": {
            "post": {
                "x-name": "/purchase/purchase",
                "description": "purchase sessions for a client",
                "operationId": "/purchase/purchase",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/purchase/updatepurchase": {
            "post": {
                "x-name": "/purchase/updatePurchase",
                "description": "update session purchase for client",
                "operationId": "/purchase/updatePurchase",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updatePurchase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/purchase/cancelpurchase": {
            "post": {
                "x-name": "/purchase/cancelPurchase",
                "description": "cancel session purchase for client",
                "operationId": "/purchase/cancelPurchase",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cancelPurchase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/purchase/fetchpurchase/{id}": {
            "get": {
                "x-name": "trainer",
                "description": "retrieve single trainer by id",
                "operationId": "purchase/fetchPurchase/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id of the session purchase",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/purchaselist/fetchpurchases/{id}": {
            "get": {
                "x-name": "/purchaseList/fetchPurchases",
                "description": "retrieve clients purchases",
                "operationId": "fetchClientsPurchases",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id of the client whose purchase we are getting",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/purchasesResponse"
                        }
                    }
                }
            }
        },
...
        "/trainer/getTrainer/{id}": {
            "get": {
                "x-name": "trainer",
                "description": "retrieve single trainer by id",
                "operationId": "trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The trainers id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "purchase": {
            "title": "purchase",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "clientId": {
                    "type": "string",
                    "format": "uuid"
                },
                "fullHour": {
                    "type": "number"
                },
                "fullHourTenPack": {
                    "type": "number"
                },
                "halfHour": {
                    "type": "number"
                },
                "halfHourTenPack": {
                    "type": "number"
                },
                "pair": {
                    "type": "number"
                },
                "pairTenPack": {
                    "type": "number"
                },
                "fullHourTotal": {
                    "type": "number"
                },
                "fullHourTenPackTotal": {
                    "type": "number"
                },
                "halfHourTotal": {
                    "type": "number"
                },
                "halfHourTenPackTotal": {
                    "type": "number"
                },
                "pairTotal": {
                    "type": "number"
                },
                "pairTenPackTotal": {
                    "type": "number"
                },
                "purchaseTotal": {
                    "type": "number"
                },
                "notes": {
                    "type": "string",
                    "x-input": "textArea"
                }
            }
        },
...
        "standardSuccessResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
        "standardFailureResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "error": {
                    "type": "boolean"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "title": "errorMessage",
                        "properties": {
                            "field": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "purchasesResponse": {
            "title": "purchasesResponse",
            "properties": {
                "trainers": {
                    "type": "array",
                    "items": {
                        "title": "purchase",
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "clientId": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "fullHour": {
                                "type": "number"
                            },
                            "fullHourTenPack": {
                                "type": "number"
                            },
                            "halfHour": {
                                "type": "number"
                            },
                            "halfHourTenPack": {
                                "type": "number"
                            },
                            "pair": {
                                "type": "number"
                            },
                            "pairTenPack": {
                                "type": "number"
                            },
                            "fullHourTotal": {
                                "type": "number"
                            },
                            "fullHourTenPackTotal": {
                                "type": "number"
                            },
                            "halfHourTotal": {
                                "type": "number"
                            },
                            "halfHourTenPackTotal": {
                                "type": "number"
                            },
                            "pairTotal": {
                                "type": "number"
                            },
                            "pairTenPackTotal": {
                                "type": "number"
                            },
                            "purchaseTotal": {
                                "type": "number"
                            },
                            "notes": {
                                "type": "string",
                                "x-input": "textArea"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {},
...
}